Abstracted and made a few improvements
